% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bellmanford}
\alias{bellmanford}
\title{Computes shortest distance from a given node to all other nodes in a given Graph.}
\usage{
bellmanford(g, s)
}
\arguments{
\item{g}{A graph list}

\item{s}{An integer indicating node}
}
\value{
If the cycle contains a negative cycle, it returns -1. Otherwise, it returns a 2 x N matrix.
The first row of the output matrix stores the shortest distance of the node \code{i} to node \code{s}.
The second row of the output matrix stores the index of predecessor node in the shortest path.
}
\description{
Computes shortest distance from a given node to all other nodes in a given Graph.
}
