% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bronars.R
\name{bronars}
\alias{bronars}
\title{bronars computes the Bronars power index corresponding to a given data set of prices and quantities}
\usage{
bronars(p, q, m)
}
\arguments{
\item{p}{T X N matrix of prices}

\item{q}{T X N matrix of quantities}

\item{m}{number of iterations}
}
\value{
Bronars power index
}
\description{
bronars computes the Bronars power index corresponding to a given data set of prices and quantities
}
