% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cycle_detection_topo}
\alias{cycle_detection_topo}
\title{Checks for a cycle in a directed graph based on topological ordering}
\usage{
cycle_detection_topo(Graph)
}
\arguments{
\item{Graph}{A graph list}
}
\value{
0 if graph is acyclic and 1 if graph is cyclic
}
\description{
Checks for a cycle in a directed graph based on topological ordering
}
