% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egarp.R
\name{egarp}
\alias{egarp}
\title{Tests consistency with the Generalized Axiom of Revealed Preference at efficiency level \code{e}}
\usage{
egarp(p, q, e)
}
\arguments{
\item{p}{T X N matrix of prices}

\item{q}{T X N matrix of quantities}

\item{e}{efficiency level}
}
\value{
1 if the data is consistent with GARP at efficiency level e, returns 0 otherwise.
}
\description{
Tests consistency with the Generalized Axiom of Revealed Preference at efficiency level \code{e}
}
\note{
T = number of observations and N = number of goods
}
