% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graph_list}
\alias{graph_list}
\title{Given data set (\code{p},\code{q}) creates appropriate graph list}
\usage{
graph_list(p, q, bound)
}
\arguments{
\item{p}{T x N matrix of prices}

\item{q}{T X N matrix of quantities}

\item{bound}{1 if upper bound is to be computed, -1 if lower bound is to be computed}
}
\value{
A graph list which is a list of \code{T} lists
}
\description{
Given data set (\code{p},\code{q}) creates appropriate graph list
}
\note{
T = number of observations and N = number of goods.
For each observation \code{i} in \code{{1,2,..,T}}, there is a list.
The list of \code{i}-th observation consists of list of 3-element vectors.
The first element stores the observation \code{(k)} such that \code{q_i R_0 q_k}.
The second element stores the weight of the edge (\code{p_iq_k-p_iq_i}) for lower bound and (\code{p_iq_i - p_iq_k}) for uppper bound.
The third element stores the budget of the edge (\code{p_iq_i}).
}
