% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indegree.R
\name{indegree}
\alias{indegree}
\title{A function to compute the number of incoming edges to each node of a given graph.}
\usage{
indegree(g)
}
\arguments{
\item{g}{A graph list}
}
\value{
A 1 X T matrix with in-degree of each node where T is the number of observations.
}
\description{
A function to compute the number of incoming edges to each node of a given graph.
}
