% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minimum_cost_time}
\alias{minimum_cost_time}
\title{Computes the minimum cost-to-time ratio of a graph}
\usage{
minimum_cost_time(Graph, low, high)
}
\arguments{
\item{Graph}{A graph list}

\item{low}{A lower bound for search value}

\item{high}{An upper bound for search value}
}
\value{
If the graph is acyclic, it returns 0. Otherwise, if the graph contains a cycle, it return the optimal ratio of minimum cost time.
}
\description{
Computes the minimum cost-to-time ratio of a graph
}
