% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpi.R
\name{mpi}
\alias{mpi}
\title{Computes bounds for the Money Pump Index}
\usage{
mpi(p, q)
}
\arguments{
\item{p}{T X N matrix of prices}

\item{q}{T X N matrix of quantities}
}
\value{
1 X 2 vector of minimum and maximum MPI of dataset (\code{p},\code{q}).
}
\description{
following Smeulders et al. (2013).
This function is meant for the analysis of prices and quantities data (e.g., of bundles of goods and corresponding prices).
}
\note{
T = number of observations and N = number of goods
}
\references{
Smeulders, Bart, Laurens Cherchye, Frits CR Spieksma, and Bram De Rock. "The money pump as a measure of revealed preference violations: A comment." Journal of Political Economy 121, no. 6 (2013): 1248-1258.
}
