% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randomq}
\alias{randomq}
\title{randomq returns a matrix of consumption bundles chosen uniformly randomly from the budgets
corresponding to a given dataset of prices and quantities}
\usage{
randomq(p, q)
}
\arguments{
\item{p}{T X N matrix of prices}

\item{q}{T X N matrix of quantities}
}
\value{
a T X N matrix of randomly choosen quantities
}
\description{
randomq returns a matrix of consumption bundles chosen uniformly randomly from the budgets
corresponding to a given dataset of prices and quantities
}
