% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefined_graph.R
\name{redefined_graph}
\alias{redefined_graph}
\title{An auxiliary function used within \code{\link{minimum_cost_time}}.
It updates the weight of each edge of a given graph.}
\usage{
redefined_graph(Graph, u)
}
\arguments{
\item{Graph}{A graph list where each edge in the graph has a cost and time weights attached to it.}

\item{u}{A scalar to update the edges of the \code{Graph}}
}
\value{
A new graph with \eqn{new cost = old cost - u*time}
}
\description{
An auxiliary function used within \code{\link{minimum_cost_time}}.
It updates the weight of each edge of a given graph.
}
