% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarp.R
\name{SARP}
\alias{SARP}
\title{Checks consistency with Strong Axiom of Revealed Preferences (SARP).}
\usage{
SARP(p, q)
}
\arguments{
\item{p}{T X N matrix of prices.}

\item{q}{T X N matrix pf quantities.}
}
\value{
\code{(passsarp,nviol)} where \code{passsarp} = 1 if data is consistent with SARP and = 0 otherwise.
\code{nviol} is the number of SARP violations.
}
\description{
Checks consistency with Strong Axiom of Revealed Preferences (SARP).
}
\note{
T = number of observations and N = number of goods
}
