% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warshall}
\alias{warshall}
\title{Algorithm to compute indirect revealed preference relations (R)}
\usage{
warshall(RO)
}
\arguments{
\item{RO}{T X T matrix of direct revealed preference relations}
}
\value{
T X T matrix of indirect revealed preference relations which is a transitive closure of \code{RO}.
}
\description{
Algorithm to compute indirect revealed preference relations (R)
}
\note{
T = number of observations
}
