% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_residual_graph.R
\name{zero_residual_graph}
\alias{zero_residual_graph}
\title{A function which makes a zero residue graph of given \code{Graph} and distances \code{d}.
This is needed to identify a zero length cycle in \code{Graph}.
If there exist a cycle in zero residual graph than there exist a zero-length cycle in original \code{Graph}.}
\usage{
zero_residual_graph(Graph, d)
}
\arguments{
\item{Graph}{A graph list}

\item{d}{2 X N matrix of shortest distance and predecessors (from node 1; computed from \code{\link{bellmanford}})}
}
\value{
Returns an updates graph
}
\description{
A function which makes a zero residue graph of given \code{Graph} and distances \code{d}.
This is needed to identify a zero length cycle in \code{Graph}.
If there exist a cycle in zero residual graph than there exist a zero-length cycle in original \code{Graph}.
}
