% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccei.R
\name{ccei}
\alias{ccei}
\title{Computes the critical cost efficiency index}
\usage{
ccei(p, q)
}
\arguments{
\item{p}{T X N matrix of prices}

\item{q}{T X N matrix of quantities}
}
\value{
the critical cost efficiency index
}
\description{
Computes the critical cost efficiency index
}
\note{
T = number of observations and N = number of goods
}
