% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garp.R
\name{garp}
\alias{garp}
\title{Tests consistency with Generalized Axiom of Revealed Preferences (GARP)}
\usage{
garp(p, q)
}
\arguments{
\item{p}{T X N matrix of prices}

\item{q}{T X N matrix of quantities}
}
\value{
\code{(passgarp, nviol)} where \code{passgarp} = 1 if the data is consistent with GARP and = 0 otherwise.
\code{nviol} is the number of GARP violations.
}
\description{
Tests consistency with Generalized Axiom of Revealed Preferences (GARP)
}
\note{
T = number of observations and N = number of goods
}
